model_training/requirements.txt
tensorflow>=2.16.0
numpy>=1.26.0
pandas>=2.0.3
scikit-learn>=1.3.0
websocket-client>=1.6.3
requests>=2.31.0
python-binance>=1.0.19
yfinance>=0.2.28
matplotlib>=3.7.2
seaborn>=0.12.2
tensorflow-model-optimization>=0.7.3
pydantic>=2.5.0

